Rename `arroyo_consumer` to `arroyo`
